refactor move findSupplier to the supplier select controller
