+++ app/pages/sections/supplier/search/controller.js
@@ -1,7 +1,4 @@
-import { getData } from 'buying-catalogue-library';
-import { getEndpoint } from '../../../../endpoints';
 import { getContext, getErrorContext } from './contextCreator';
-import { logger } from '../../../../logger';
 
 export const getSupplierSearchPageContext = async ({ orderId }) => (
   getContext({ orderId })
@@ -22,11 +19,3 @@ export const validateSupplierSearchForm = ({ data }) => {
 };
 
 export const getSupplierSearchPageErrorContext = params => getErrorContext(params);
-
-export const findSuppliers = async ({ supplierNameToFind, accessToken }) => {
-  const endpoint = getEndpoint({ endpointLocator: 'getSearchSuppliers', options: { supplierNameToFind } });
-  const suppliersFound = await getData({ endpoint, accessToken, logger });
-  logger.info(`Searching for "${supplierNameToFind}" returned ${suppliersFound.length} supplier`);
-
-  return suppliersFound;
-};
