+++ app/pages/sections/supplier/search/controller.test.js
@@ -1,6 +1,6 @@
 import { getData } from 'buying-catalogue-library';
 import * as contextCreator from './contextCreator';
-import { getSupplierSearchPageContext, validateSupplierSearchForm, findSuppliers } from './controller';
+import { getSupplierSearchPageContext, validateSupplierSearchForm } from './controller';
 import { solutionsApiUrl } from '../../../../config';
 import { logger } from '../../../../logger';
 
@@ -71,25 +71,6 @@ describe('supplier search controller', () => {
         const response = validateSupplierSearchForm({ data });
 
         expect(response.errors).toEqual(expectedValidationErrors);
-      });
-    });
-  });
-
-  describe('findSuppliers', () => {
-    afterEach(() => {
-      getData.mockReset();
-    });
-
-    it('should call getData once with the correct params', async () => {
-      getData
-        .mockResolvedValueOnce({ data: [] });
-
-      await findSuppliers({ supplierNameToFind: 'some-supp', accessToken: 'access_token' });
-      expect(getData.mock.calls.length).toEqual(1);
-      expect(getData).toHaveBeenCalledWith({
-        endpoint: `${solutionsApiUrl}/api/v1/suppliers?name=some-supp`,
-        accessToken: 'access_token',
-        logger,
       });
     });
   });
